% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annHeatmap.R
\name{getLeaves}
\alias{getLeaves}
\title{Undocumented functions}
\usage{
getLeaves(x)
}
\description{
These functions are currently undocumented. Please refer to the source and
source comments if you feel you need to use them.
}
\keyword{internal}
