\name{Undocumented functions}
\alias{getLeaves}
\alias{picketPlotControl}
\title{Undocumented functions}
\description{
\code{getLeaves(x)}

\code{picketPlotControl()}

These functions are currently undocumented. Please refer to the source and source comments if you feel you need to use them.
}
\keyword{internal}% at least one, from doc/KEYWORDS

