% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annHeatmap.R
\name{print.annHeatmap}
\alias{print.annHeatmap}
\title{Printing information about annotated heatmaps}
\usage{
\method{print}{annHeatmap}(x, ...)
}
\arguments{
\item{x}{an object of class \code{annHeatmap}}

\item{\dots}{extra arguments, currently ignored}
}
\value{
\code{x} is returned invisibly
}
\description{
Printing method for annotated heatmaps
}
\details{
A very simple printing method, displaying a minimum of information about
dendrograms and annotation
}
\examples{

    set.seed(219)
    mat = matrix(rnorm(100), ncol=5)
    ann = data.frame(Class=c("A","A","B","A","B"))
    map1 = annHeatmap(mat, ann)
    map1

}
\seealso{
\code{\link{annHeatmap}}, \code{\link{annHeatmap2}},
\code{\link{plot.annHeatmap}}
}
\keyword{hplot}
